% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorsDataBGD}
\alias{indicatorsDataBGD}
\title{WASH Survey Indicators Data for Dhaka, Bangladesh}
\format{A data frame with 162 variables and 1282 entries}
\usage{
indicatorsDataBGD
}
\description{
WASH survey indicators data calculated from survey raw data
}
\keyword{datasets}
