% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popBGD}
\alias{popBGD}
\title{Population Data - Dhaka, Bangladesh}
\format{A data frame with four variables: \code{psu}, \code{zone},
  \code{type}, \code{pop}.}
\usage{
popBGD
}
\description{
Grid-based population of Dhaka, Bangladesh
}
\keyword{datasets}
