% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixBGD}
\alias{ppiMatrixBGD}
\title{PPI Look-up Table for Bangladesh}
\format{A data frame with 10 variables and 101 entries}
\usage{
ppiMatrixBGD
}
\description{
PPI look-up table for Bangladesh to calculate PPI score
}
\keyword{datasets}
