% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{washdata}
\alias{washdata}
\alias{washdata-package}
\title{Urban Water and Sanitation Survey Dataset}
\description{
See the README at
\href{https://github.com/validmeasures/washdata#readme}{GitHub}
}
