% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popBGD}
\alias{popBGD}
\title{Population Data for Dhaka, Bangladesh}
\format{
A data frame with four variables:
\describe{
\item{\code{psu}}{primary sampling unit (PSU) ID}
\item{\code{zone}}{survey enumeration area}
\item{\code{type}}{slum (1) or non-slum (2)}
\item{\code{pop}}{population}
}
}
\usage{
popBGD
}
\description{
Grid-based population of Dhaka, Bangladesh
}
\examples{
popBGD
}
\keyword{datasets}
