% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveyDataBGD}
\alias{surveyDataBGD}
\title{WASH Survey Raw Data for Dhaka, Bangladesh}
\format{A data frame with 218 variables and 1282 entries}
\usage{
surveyDataBGD
}
\description{
WASH survey raw data collected by WSUP in Dhaka, Bangladesh
}
\keyword{datasets}
