% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveyDataBGD}
\alias{surveyDataBGD}
\title{WASH Survey Data - Dhaka, Bangladesh}
\format{An object of class \code{data.frame} with 1282 rows and 218 columns.}
\usage{
surveyDataBGD
}
\description{
WASH survey data collected by WSUP in Dhaka, Bangladesh using
}
\keyword{datasets}
